% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impedance_calculate.R
\name{impedance_calculate}
\alias{impedance_calculate}
\title{calulate impedance from CA experiment}
\usage{
impedance_calculate(ca_data, Amplitude = 0.01, ratio = 10)
}
\arguments{
\item{ca_data}{CA experimental data for calculate}

\item{Amplitude}{Amplitude used in CA, in unit of Voltage}

\item{ratio}{ratio for high frequency data cutoff}
}
\value{

}
\description{
calulate impedance from CA experiment
}
\examples{

}
