% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ec.R
\name{list_ec}
\alias{list_ec}
\title{list electrochemical data files using regex}
\usage{
list_ec(dir, type)
}
\arguments{
\item{dir}{directory where ec data in}

\item{type}{a character, what data to list}
}
\value{

}
\description{
list electrochemical data files using regex
}
