% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tafel.R
\name{list_tafel}
\alias{list_tafel}
\title{list cv data files using regex}
\usage{
list_tafel(dir)
}
\arguments{
\item{dir}{a string, directory where cv data in}
}
\value{

}
\description{
list cv data files using regex
}
