% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mav.R
\name{mav}
\alias{mav}
\title{Moving average smooth with varying window sizes}
\usage{
mav(x)
}
\arguments{
\item{x}{A column of data for processing}
}
\value{

}
\description{
Moving average smooth with varying window sizes
}
