% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Calculate and generate 2D COS with corr2D and ggplot2}
\usage{
plot_corr(x, type = "Re")
}
\arguments{
\item{x}{input matrix for 2D COS calculation}

\item{type}{Re for Synchronous and Im for asynchronous 2D correlation spectra}
}
\value{

}
\description{
Calculate and generate 2D COS with corr2D and ggplot2
}
