% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv.R
\name{plot_cv}
\alias{plot_cv}
\title{plot cv}
\usage{
plot_cv(data, color = "Sample", asp = 1, ...)
}
\arguments{
\item{data}{data used for plot}

\item{color}{color for different sample}

\item{asp}{aspect ratio of plot}

\item{...}{}
}
\value{

}
\description{
plot cv
}
