% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tafel.R
\name{plot_tafel}
\alias{plot_tafel}
\title{plot tafel data}
\usage{
plot_tafel(data, color = "Sample", asp = 1)
}
\arguments{
\item{data}{data used for plot}

\item{color}{color for different samples}

\item{asp}{aspect ratio of plot}
}
\value{

}
\description{
plot tafel data
}
