% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_theme.R
\name{pub_theme}
\alias{pub_theme}
\title{Customed ggplot theme for publication}
\usage{
pub_theme(asp = 1)
}
\arguments{
\item{asp}{control the aspect ratio of plot, default to 1.}
}
\value{

}
\description{
Customed ggplot theme for publication
}
