% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ec.R
\name{read_ec}
\alias{read_ec}
\title{Read electrochemical data}
\usage{
read_ec(dir, type)
}
\arguments{
\item{dir}{directory where ec data in}

\item{type}{a character, what data to read}
}
\value{

}
\description{
Read electrochemical data
}
