% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eis.R
\name{plot_eis}
\alias{plot_eis}
\title{plot EIS data}
\usage{
plot_eis(data, type = "N", size = 1, asp = 1, color = "Sample")
}
\arguments{
\item{data}{EIS data for plot}

\item{type}{plot type desired, N for Nyquist plot and B for Bode plot, default as Nyquist}

\item{size}{point size, default as 1}

\item{asp}{aspect ratio of plot}

\item{color}{point color, default as blue}
}
\value{

}
\description{
plot EIS data
}
